Adapt HeatPump and HeatExchanger to mass balance HeatCarrier
The more verbose structure of the HeatPump is equivalent and should be
easier to understand compared to the 4-input, 2-output version with just
one Converter. At the same time, it reduces the number of Flows that
connect to outside the HP and simplifies data analysis as there is now
one Flow representing all heat coming from and going to each temperature
level (at the hot/cold side, respectively).